requirements.txt
gradio
openai
anthropic
google-genai
openpyxl
